DPL 1316 Add module for paths
Currently we have a lot of hard-coded paths and code for extracting
information from paths.

We want to extract this into one place in the base container layer, so that we
don't duplicate logic between containers. Doing so will allow us to extend or
modify our s3 bucket structure without having to change the whole codebase.

As a first step, this module introduces a new API for constructing S3 paths
and Athena table names, and parsing them to extract table name, timestamp etc.

fixup - rename