Remove special path logic from infer_glue_schema
Note there is a slight change of behaviour here.

Before, the database name was based on a parameter, and when called with raw
data, it was being set to the curated database name, (i.e. the one derived from
the data product name). Now, we ask the data_product_config for both the
database name and the table name, which will return a value like
('data_products_raw', 'adjudications_raw'). This behaviour is more consistent,
but we'll need to make sure we handle it correctly in the calling code.