Remove special path logic from create_curated_athena_table
I've changed the create_curated_athena_table function to accept all the boto
clients as parameters, to make it easier to test.

We're now unit testing some (but not all) of the functionality of
create_curated_athena_table:

- creating the glue db if it is missing
- checking if any partition files exist
- formatting athena queries
- the function doesn't error if partition files already exist
- the function doesn't error if the glue table already exists

Unfortunately we can't test executing the athena queries,
since moto does not support this.