Generate pytest path filter config on the fly
We already have a script to generate this, but currently we check in the result
instead of running it as part of the CI. This means every time we add a new
python container, we have to rerun it, and the changes have to be reviewed by
ministryofjustice/data-platform-core-infra.

Thanks to @murdo-moj for the suggestion.