Refactor athena load lambda to use shared paths module (#1469)
* Add unit tests to daap-athena-load

Make sure that

- The tests run in CI
- The linter doesn't bug us about repeated code in tests

* Add __eq__ method

* Remove special path logic from infer_glue_schema

Note there is a slight change of behaviour here.

Before, the database name was based on a parameter, and when called with raw
data, it was being set to the curated database name, (i.e. the one derived from
the data product name). Now, we ask the data_product_config for both the
database name and the table name, which will return a value like
('data_products_raw', 'adjudications_raw'). This behaviour is more consistent,
but we'll need to make sure we handle it correctly in the calling code.

* Remove special path logic from create_raw_athena_table

Note that now `infer_metadata_schema` returns the correct database name for raw
data, we can pull the value straight from the metadata.

* Remove special path logic from create_curated_athena_table

I've changed the create_curated_athena_table function to accept all the boto
clients as parameters, to make it easier to test.

We're now unit testing some (but not all) of the functionality of
create_curated_athena_table:

- creating the glue db if it is missing
- checking if any partition files exist
- formatting athena queries
- the function doesn't error if partition files already exist
- the function doesn't error if the glue table already exists

Unfortunately we can't test executing the athena queries,
since moto does not support this.

* Update athena_load_handler to pass DataProductConfig

This completes the work to remove all code to do with paths and table name
conventions from this lambda.