Detect and reject early unprocessable CSV formats (#2162)
* Detect and reject early unprocessable CSV formats

The athena load step cannot process certain kinds of CSVs due to
limitations in the OpenCSVSerde used by Athena.

Specifically, this does not support embedded line breaks in CSV files,
even though this is valid CSV, and can be produced by a postgres CSV
export.

See
https://docs.aws.amazon.com/athena/latest/ug/csv-serde.html
https://repost.aws/questions/QUwL_lVp85TPe-i3VP6jcU_A/athena-not-able-to-read-multi-line-text-in-csv-fields

This commit detects this happening, and fails the ingestion.

* Reduce log level when table is first created

This scenario is normal when first uploading to a data product. Making
this `info` so it doesn't clutter the error logs.