Directly define metadata schema in the OpenAPI specification
Previuosly we were pointing to our internal metadata spec that includes
platform generated metadata.

To avoid confusion, we want to distinguish the fields that the user can
actually specify in register/update requests.

Note that internally, we will still validate against the json schema
within the modernisation-platform-environments repo, so we will need to
maintain both in parallel.

In the case of table schemas, everything is settable by the user, so
the documentation is ok as it is.