Ensure subtypes conform to base type
The aim here is to hide catalogue specific details from the user of
this library.

This means we should expect to be able to type any value to
BaseCatalogueClient and then be able to substitute any of the concrete
classes without breaking things. (See Liskov's substitution principle)

This cannot be the case if the base class has methods which accept *any*
arguments, and the concrete implementations only handle specific
arguments.

To fix this I've modified the abstract methods to specify the arguments.

Since the schema_fqn, database_fqn, etc arguments reflected the
OpenMetadata graph structure, I've replaced this with a generic
"DataLocation" argument, which does not enforce a particularly
hierarchy.