Add a method for search
This method accepts a query and pagination variables, and returns
the total number of results plus the results to display on the
current page.

For each search result, return:

- the result type, DataSet or DataProduct
- ID, name and description
- tags
- a dictionary of additional metadata fields (exactly what this will
  contain will depend on what tests well with users)
- information about the properties which matched the search query
- the time the metadata was last updated

The raw responses from Datahub are logged at DEBUG level.

For now I've excluded filters, facets, and sorting but these are
supported by the underlying API.

See
- https://datahubproject.io/docs/graphql/inputObjects#facetfilterinput
- https://datahubproject.io/docs/graphql/objects#facetmetadata
- https://datahubproject.io/docs/graphql/inputObjects#searchsortinput