Allow the search response to contain facet information
Facets are the dynamic search filters that show how the search results
break down across different dimensions. Returning these allows us
to display only options that are relevant to the current result set
and indicate the number of results matching each option.

For most filters, we are expected to pass a URN as the value, so
this is also the easiest way for the frontend to figure out what
the possible values are.