Replace data-catalogue with a new repo
- Rename data-catalogue back to find-moj-data
- Create a new data-catalogue repo

Since find-moj-data creates a build artefact, we want it to have its own
repo.

However we also want a "top level" repo that links out to the various
artefacts. It is this repo that would take over the repository
terraforming if we decouple from analytical platform. Right now, we will
use it for the helm deployment of datahub itself, and any documentation
or tickets related to the overall data catalogue project.