Return created and modified timestamps
Return created and modified timestamps for datasets and search results.

Previously we were returning the last ingested timestamp for search
results, which is not exactly what we want to present to users. It gives
an indication of whether metadata may be stale or not, but we shouldn't
present that as when the metadata was last modified, or the user will
conclude that there has been some change that never happened.

Returning this for datasets as well means we can display it on the
details page.