Add missing metadata when fetching datasets (#3722)
* Test tags in dataset response

* Return native types if available

When presenting datasets in our UI, use the native data types if we have
them. We suspect these will be more useful to the users.

* Return qualifiedName in datasets if available

* Return created and modified timestamps

Return created and modified timestamps for datasets and search results.

Previously we were returning the last ingested timestamp for search
results, which is not exactly what we want to present to users. It gives
an indication of whether metadata may be stale or not, but we shouldn't
present that as when the metadata was last modified, or the user will
conclude that there has been some change that never happened.

Returning this for datasets as well means we can display it on the
details page.

* Bump version to 0.23.0

* Lint

* Fall back to top level name attribute if missing from properties

According to https://datahubproject.io/docs/graphql/objects/#dataset
name is not supposed to be used anymore, but it seems like the dbt
ingestion is populating this instead of properties.name.