Simplify types
Custom properties values are always String in datahub, so using string
throughout will make our lives easier. We can use empty string to
indicate missing values here.
See https://datahubproject.io/docs/graphql/objects/#custompropertiesentry

dpia_required is a special case - it's still a string in datahub but we
want it to be a boolean when we use it - since "False" is truthy
in python.

When setting custom properties, we take the string representation. This
will serialize booleans to 'True' and 'False' which will fail pydantic's
validation, so we need to explicitly convert this back to a boolean.
https://docs.pydantic.dev/2.0/usage/types/booleans/

If dpia_required is unset, we should just pass empty string, rather than
"None".